refactor: split type declarations into multiple files
