+++ nominal/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@coderspirit/nominal",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "Powerful nominal types for your project",
   "main": "./dist/index.js",
   "keywords": [
