+++ nominal/src/Tainted.ts
@@ -1,0 +1,4 @@
+import { GenericTainted } from 'GenericTainted'
+import { TaintSymbolType } from 'internal/Symbols'
+
+export type Tainted<BaseType> = GenericTainted<BaseType, TaintSymbolType>
