+++ nominal/src/Untainted.ts
@@ -1,0 +1,4 @@
+import { GenericUntainted } from './GenericUntainted'
+import { TaintSymbolType } from './internal/Symbols'
+
+export type Untainted<BaseType> = GenericUntainted<BaseType, TaintSymbolType>
