+++ nominal/src/internal/Symbols.ts
@@ -1,0 +1,2 @@
+export const TaintSymbol: unique symbol = Symbol('TaintSymbol')
+export type TaintSymbolType = typeof TaintSymbol
